Update post components folder path
