+++ src/components/post/detail/CommentSection.tsx
@@ -1,0 +1,10 @@
+import { Comment } from "~/apollo/generated/graphql";
+
+type Props = {
+  comment: Comment;
+};
+function CommentSection({ comment }: Props) {
+  return <div>{comment.id}</div>;
+}
+
+export default CommentSection;
